+++ UMC-iOS/UMC-iOS/Data/Network/NetworkManager/BoardNetwork.swift
@@ -7,57 +7,6 @@
 
 import Foundation
 
-extension NSMutableData {
-    func appendString(_ string: String) {
-        let data = string.data(using: String.Encoding.utf8, allowLossyConversion: true)
-        append(data!)
-    }
-}
-
-extension Data {
-    static func createJsonStringData(boundary: String, data: Data) -> Data {
-        let body = NSMutableData()
-        let boundaryPrefix = "--\(boundary)\r\n"
-
-        body.appendString(boundaryPrefix)
-        body.appendString("Content-Disposition: form-data; name=\"request\"\r\n")
-        body.appendString("Content-Type: application/json\r\n\r\n")
-        body.append(data)
-        body.appendString("\r\n")
-//        body.appendString("--".appending(boundary.appending("--")))
-
-        return body as Data
-    }
-    
-    static func createFileData(boundary: String, data: Data, mimeType: String, fileName: String) -> Data {
-        let body = NSMutableData()
-        let boundaryPrefix = "--\(boundary)\r\n"
-
-        body.appendString(boundaryPrefix)
-        body.appendString("Content-Disposition: form-data; name=\"file\"; filename=\"\(fileName).\(mimeType)\"\r\n")
-        body.appendString("Content-Type: application/\(mimeType)\r\n\r\n")
-        body.append(data)
-        body.appendString("\r\n")
-//        body.appendString("--".appending(boundary.appending("--")))
-
-        return body as Data
-    }
-    
-    func combineData(datas: [Data]) -> Data {
-        let combinedData = NSMutableData()
-        
-        for data in datas {
-            combinedData.append(data)
-        }
-        return combinedData as Data
-    }
-}
-
-struct FileInfo {
-    var data: Data
-    var fileName: String
-    var mimeType: String
-}
 
 class BoardNetwork: ObservableObject {
     
@@ -229,5 +178,13 @@ class BoardNetwork: ObservableObject {
         print(boardId)
         
         return boardId
+    }
+}
+
+extension BoardNetwork {
+    struct FileInfo {
+        var data: Data
+        var fileName: String
+        var mimeType: String
     }
 }
