+++ UMC-iOS/UMC-iOS/Extension/NSMutableData.swift
@@ -1,0 +1,15 @@
+//
+//  NSMutableData.swift
+//  UMC-iOS
+//
+//  Created by Kyungsoo Lee on 2/15/24.
+//
+
+import Foundation
+
+extension NSMutableData {
+    func appendString(_ string: String) {
+        let data = string.data(using: String.Encoding.utf8, allowLossyConversion: true)
+        append(data!)
+    }
+}
